Refine radio initialization flow
